+++ core/src/main/kotlin/config/ApiWebClientConfig.kt
@@ -3,35 +3,32 @@ package com.wafflestudio.snu4t.config
 import io.netty.channel.ChannelOption
 import io.netty.handler.timeout.ReadTimeoutHandler
 import io.netty.handler.timeout.WriteTimeoutHandler
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Profile
 import org.springframework.http.client.reactive.ReactorClientHttpConnector
+import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
 import reactor.netty.Connection
 import reactor.netty.http.client.HttpClient
-import reactor.netty.resources.ConnectionProvider
-import java.time.Duration
 import java.util.concurrent.TimeUnit
 
 @Configuration
+@EnableConfigurationProperties(ApiConfigs::class)
 class ApiWebClientConfig(
     private val apiConfigs: ApiConfigs
 ) {
     @Bean
+    @Profile("!test")
     fun snuttevServer(): SnuttEvWebClient {
         return create(apiConfigs.server["snuttev"]!!).let(::SnuttEvWebClient)
     }
 
     private fun create(apiConfig: ApiConfig): WebClient {
         val connector = ReactorClientHttpConnector(
-            HttpClient.create(
-                ConnectionProvider.builder(apiConfig.baseUrl)
-                    .maxConnections(500)
-                    .pendingAcquireTimeout(Duration.ofMillis(100))
-                    .maxLifeTime(Duration.ofSeconds(600))
-                    .maxIdleTime(Duration.ofSeconds(60))
-                    .build()
-            )
+            HttpClient.create()
                 .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, apiConfig.connectTimeout) // Connection Timeout
                 .doOnConnected { conn: Connection ->
                     conn
@@ -41,6 +38,18 @@ class ApiWebClientConfig(
         )
         return WebClient.builder().clientConnector(connector).baseUrl(apiConfig.baseUrl).build()
     }
+}
+
+@Component
+@ConfigurationProperties(prefix = "api")
+class ApiConfigs {
+    var server: Map<String, ApiConfig> = hashMapOf()
+}
+
+class ApiConfig {
+    var connectTimeout: Int = 800
+    var readTimeout: Long = 1000
+    lateinit var baseUrl: String
 }
 
 class SnuttEvWebClient(webClient: WebClient) : WebClient by webClient
