+++ core/src/testFixtures/kotlin/mock/api/MockApiWebClientConfig.kt
@@ -1,0 +1,14 @@
+package mock.api
+
+import com.wafflestudio.snu4t.config.SnuttEvWebClient
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Primary
+import org.springframework.stereotype.Component
+import org.springframework.web.reactive.function.client.WebClient
+
+@Component
+class MockApiWebClientConfig {
+    @Bean
+    @Primary
+    fun snuttevServer(): SnuttEvWebClient = SnuttEvWebClient(WebClient.create())
+}
